api/requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
sqlalchemy==2.0.1
psycopg2-binary==2.9.5
alembic==1.9.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
